Initialize quiz-server repository
This commit starts a new repository featuring a minimal
backend service (DB and server) that partially emulates
an existing quiz application.

The application (server) and migration code uses Clojure.
Configuration of the supporting database is achieved with
docker-compose, aiming to aid in portability across
systems.

A small set of behavior tests are included. Documentation
emphasizing choices and describing the server api and
architecture are included in a dedicated directory.

Ongoing issues with including dates in timestamps are noted
but unresolved.